requirements.txt
sympy==1.13.3
Pillow==11.1.0
pytesseract==0.3.13
